requirements.txt
newspaper3k 
streamlit 
openai
pandas
requests
python-dotenv